Split test-lib into three libraries
test-lib contained two kinds of infrastructure that shouldn't co-exist
in the same library:
- Generic infrastructure that would be used by all test suites.
- Suite-specific infrastructure (i.e. the message definition) that is
  only useful to the specific test suite in this repository.

This isn't that relevant right now, as there is just one test suite, and
everything lives in the same repository anyway, but the plan for this
project is not just to build a test suite, it's to build a general test
framework. So long-term, we need to split generic and suite-specific
code into different packages, and now seemed as good a time to do it as
any.

So why did this result in three libraries? As it turned out, after I was
done cleaning up, there was no shared code left between the support code
for the firmware and host sides (or rather, that shared code lives in
3rd-party libraries like Serde and Postcard). The needs of both sides
are diametrically opposed anyway (requiring the two features in
test-lib), and splitting them simplified things a bit.