Share one target directory between all projects
I have at several points attempted to add everything here to a single
workspace, but this didn't work out. Workspaces generally assume some
uniformity, but the firmware- and host-specific crates are a bit too
different in this regard.

This commit adds configuration to all crates that defines a single,
shared target directory, which is one of the features that a workspace
provides. This should reduce the time required for a clean build, which
is especially relevent for CI.