Add API for test assistant
This duplicates some code from `Target` (the methods that receive
messages are nearly identical). The different message type makes it not
quite straight-forward to use the same implementation. And all the non-
straight-forward solutions I can think of seem like overengineering,
given that I don't fully know how both structs will evolve.

This is not ideal, but I think it's good enough for now. It's better to
revisit this later, when everything has become a bit more fleshed out.