Simplify error handling for assistant API
This merges all the previously separate error into one enum, making it
easier to handle those errors. I am about to move the assistant API to
host-lib, so it can be re-used in the STM32L4 Test Suite. This change
will reduce code duplication between the test suites.

Reducing the granularity of error handling like this can be a
disadvantage, as it might force a caller to write error handling code
for errors that can't actually happen for a given method call. This
isn't a problem in this case, as no error handling like that happens in
a test case. The only point of the granularity was to make it obvious
where an error originated from, and this characteristic is kept here.